fix(routes): move detail flows to root stack
Event and moment detail pages now open as full-page screens above the tabs.

Create and edit routes stay in the same navigator so transitions and back behavior stay consistent.
