+++ app/(tabs)/events/_layout.tsx
@@ -10,16 +10,6 @@ export default function EventsLayout() {
       }}
     >
       <Stack.Screen name="index" />
-      <Stack.Screen
-        name="new"
-        options={{
-          presentation: 'modal',
-        }}
-      />
-      <Stack.Screen
-        name="[id]"
-        options={{ animation: 'slide_from_bottom', presentation: 'modal' }}
-      />
     </Stack>
   );
 }
